Update project dependencies moves IdentityServer to be on the UserModule Infraestructure
