+++ src/Modules/UserAccess/Application/CompanyName.MyMeetings.Modules.UserAccess.Application.csproj
@@ -2,7 +2,4 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Meetings\IntegrationEvents\CompanyName.MyMeetings.Modules.Meetings.IntegrationEvents.csproj" />
   </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="IdentityServer4" Version="4.1.2" />
-  </ItemGroup>
 </Project>
