+++ src/Modules/UserAccess/Application/Contracts/CustomClaimTypes.cs
@@ -1,9 +1,9 @@
 ﻿namespace CompanyName.MyMeetings.Modules.UserAccess.Application.Contracts
 {
-    internal class CustomClaimTypes
+    public class CustomClaimTypes
     {
-        internal const string Roles = "roles";
-        internal const string Email = "email";
-        internal const string Name = "name";
+        public const string Roles = "roles";
+        public const string Email = "email";
+        public const string Name = "name";
     }
 }
